:truck: Rename InputDevice to InputMethod
