+++ src/renderer/menu/input-methods/gesture-detector.ts
